[MOBILE-7164] Updated project to accommodate Xcode 12
- Added 'EXCLUDED_ARCHS' for simulator builds because xcode is adding arm64 archs to simulator builds (this might be a bug in Xcode12?)
- Change minimum deployment target for iOS 8.0 -> 9.0
- updated modular-crash-reporter-ios to accommodate deprecations in iOS 9.0.
- updated podspec to include EXCLUDED_ARCHS
- various other build warnings fixed.

+++ Agent/Analytics/NRMANetworkResponseData.mm
@@ -20,7 +20,7 @@
                     responseTime:(double)timeInSeconds {
     self = [super init];
     if(self) {
-        wrappedNetworkResponseData = new NewRelic::NetworkResponseData(statusCode, bytesReceived, timeInSeconds);
+        wrappedNetworkResponseData = new NewRelic::NetworkResponseData((int)statusCode, (unsigned int)bytesReceived, timeInSeconds);
         if(!wrappedNetworkResponseData)
             self = nil;
     }
@@ -33,8 +33,8 @@
         networkErrorMessage:(NSString*)errorMessage {
      self = [super init];
      if(self) {
-         wrappedNetworkResponseData = new NewRelic::NetworkResponseData(networkErrorCode,
-                                                                        bytesReceived,
+         wrappedNetworkResponseData = new NewRelic::NetworkResponseData((int)networkErrorCode,
+                                                                        (unsigned int)bytesReceived,
                                                                         timeInSeconds,
                                                                         errorMessage.UTF8String);
          if(!wrappedNetworkResponseData)
@@ -57,8 +57,8 @@
             encodedResponseBody = [NRMABase64 encodeFromData:[responseBody dataUsingEncoding:NSUTF8StringEncoding]];
         }
         
-        wrappedNetworkResponseData = new NewRelic::NetworkResponseData(statusCode,
-                                                                       bytesReceived,
+        wrappedNetworkResponseData = new NewRelic::NetworkResponseData((int)statusCode,
+                                                                       (unsigned int)bytesReceived,
                                                                        timeInSeconds,
                                                                        errorMessage.UTF8String,
                                                                        encodedResponseBody.UTF8String,
